move test file into top level component
